🚚 Change structure of states
